Target contour shape filtering. New database with 166 real images.
Shape filtering with computation of aspect ratio and circularity of
target contour (obtained with the homograpy).
These measures are compared with a database of allowed shapes and
filtered if their are not in the allowed range.
Refactoring of findHomograpy function from OpenCV to allow the fine
tunning of the ransac method confidence and maximum number of
iterations.